split ControlPanel to fix its display
Control panel display was broken since SailfishOS 3.
It seemed to be due to DockedPanel having states (not sure if they were
there in previous versions) which probably conflicted with our own
playing/pause states.

it is now split between an Item containing the states, info,... and the
panel keeping the UI part